% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_example.R
\name{area_example}
\alias{area_example}
\title{Get Path to Example Image}
\usage{
area_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If NULL, the example files will be listed.}
}
\description{
LAV comes bundled with a number of sample files in its inst/extdata directory. This function make them easy to access.
}
\examples{
area_example()
area_example("raw/img5.jpg")
}
