% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess.R
\name{assess}
\alias{assess}
\title{Assess Leaf Area for a Single Image}
\usage{
assess(
  img,
  threshold = 120,
  cut_off = 10000,
  output_dir = NULL,
  combine = FALSE,
  res = 300
)
}
\arguments{
\item{img}{The path to the image on which you want to assess the leaf area}

\item{threshold}{A value between 0 (black) and 255 (white) for classification of background and leaf pixels. Default = 120}

\item{cut_off}{Clusters with fewer pixels than this value will be discarded. Default is 10000, which is about 3.3 pixels x 3.3 pixels, in a 300 dots per inch image.}

\item{output_dir}{The directory where to save the processed images. Respects tilde expansion. By default, processed images are not saved. This must not be a directory that already exists, so that existing files are not over-written.}

\item{combine}{If true the total area will be returned; otherwise each segment will be returned separately}

\item{res}{Image resolution, in dots per inch (DPI); if False the resolution will be read from the exif tag. Default is 300 DPI.}
}
\value{
The assessed leaf area for a single image or entire folder of images, in cm^2 is returned as a \code{data.frame}.
}
\description{
This function assesses the area of leaves taken from a single image. Ideally, the image will be obtained from a flatbed scanner, and will include resolution information in the exif metadata. Images derived from cameras can also be used, but in that case, you will probably need to supply the resolution explictly, using the argument res.
}
\examples{
img <- area_example("prepared/img1.jpg")
output_dir <- area_example("processed")


# If the argument output_dir is omitted,  do not save  processed images
assess(img, res = 400, combine = TRUE)
\dontrun{
assess(img, output_dir = output_dir, res = 400, combine = F)
}
}
